Extract episode creation from database to the model itself
